[REF] Move pathchers to hooks.py file
